lpl/requirements.txt
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.1
seaborn==0.12.2
PILLOW==9.5.0
scikit-learn==1.2.2
wandb==0.15.4
tqdm==4.65.0
